[Refactor] IconButton에서 SelectedIconButton으로 네이밍 변경 (#46)
