+++ src/shared/ui/button/index.ts
@@ -1,2 +1,2 @@
 export { default as Button } from "./Button";
-export { default as IconButton } from "./IconButton";
+export { default as SelectedIconButton } from "./SelectedIconButton";
